backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
anthropic==0.18.1
sqlalchemy==2.0.25
pydantic==2.6.1
python-multipart==0.0.9
mangum==0.17.0
psycopg2-binary==2.9.9  # PostgreSQL adapter 